Releases version 51.0.0 with licensing update (#67)
* Increments major version to 51.0.0

Bumps the major version to 51.0.0, reflecting significant new features and updates.

Adds new "Flash" cartridge.

Updates README to reflect new version and features.

* Clarifies licensing and adds feature info

Elaborates on the licensing terms for different aspects of the software.

Updates the README to highlight new features and provides a visual aid
with an image.

* Updates free cartridge links

Updates links to free cartridges to point to their new locations in the repository.

This improves discoverability and ensures users can easily access the cartridge resources.

* Adds a "Happy Days" advertisement

Adds a "Happy Days" advertisement to the README file.

Removes a placeholder image.