Updates site branding and assets
Updates the site's favicon and main icon.

Replaces the old icon with a new "blokey" icon in the markdown configuration file.

Adds new favicon and iOS icons in both PNG and Sketch formats.
Removes the older version of the iOS icon.

Replaces the old favicon SVG with a new one.