Enhances Fallmanager with Raw Text Extraction (#132)
* Enables raw text extraction from uploaded files

Adds functionality to extract raw text from uploaded files using PDF.co API.
This allows for better indexing and searching of file content.
The extracted text is stored in Firestore and displayed in the file edit view.
Also removes the old, unused RIP route.

* Rearranges file details and thumbnail display.

Moves the file details and thumbnail display to improve the component's structure and readability.
The thumbnail is displayed with a loading indicator and a link to the downloadable file.
The file details are displayed in a separate grid.

* Enhances Fallmanager file management UI/UX

Refines the Fallmanager UI components for better user experience.

- Updates the Files component to remove file type and improve time display.
- Improves thumbnail generation and text extraction feedback in FileEdit.
- Adjusts API error responses for clarity.

* Adds CV cartridge documentation

Provides a comprehensive README for the CV cartridge, detailing its purpose, update process, branching strategy, and deployment.

This document guides developers on how to modify and deploy changes to the CV content effectively.

* Updates markdown formatting in README

Corrects the markdown syntax in the README file for the "Promote" action, changing it from bold to italics for better readability.

* Updates title to be more descriptive

Updates the CV page title to "Fullstack JavaScript Developer"
to provide a clearer indication of the content.

* Updates global navigation title

Updates the title of a navigation item in the global navigation data.

The title "CV" is updated to "Fullstack JavaScript Developer" for better clarity and relevance.