Adds Flickr recache functionality to the UI
Enables users to manually trigger a refresh of the latest Flickr photos from the UI.

This commit introduces a button within the Flickr component that dispatches an action to call the Flickr API and update the cached data.
A feedback message is displayed to inform the user about the ongoing process.

The global nav images are also updated.