Adds Flickr photo cartridge with latest photos (#139)
* Adds Flickr Photo Cartridge page

Adds a dedicated page for the Flickr Photo Cartridge.

Updates the navigation and core component to reflect the new page's location.

Creates new markdown content and updates existing content to improve descriptions.

* Adds Flickr latest photos functionality

Adds a new Flickr component to display the latest photos from a Flickr account.

This includes creating an API endpoint to fetch the latest photos,
a new component to display the photos,
and updates to the Flickr cartridge to integrate the new component.

* Fetches and stores latest Flickr photos

Implements an API endpoint to fetch the latest photos from Flickr, including metadata and different sizes.
Stores the fetched photos in Firebase, allowing retrieval for display.
Replaces the previous implementation using flickr.photosets.getList with flickr.people.getPublicPhotos for more accurate results.
Fixes data persisting issues in Firebase by setting `ignoreUndefinedProperties` to true.

* Enhances CV analysis and UI improvements

Refactors the CV analysis feature to provide better user guidance and examples.

This commit improves the user experience by adding example links for job descriptions, enhancing button appearance, and adjusting the layout for better readability across devices. It also removes unused elements, and pulls in imagery from Flickr to improve the appearance of the site.

* Adds Flickr recache functionality to the UI

Enables users to manually trigger a refresh of the latest Flickr photos from the UI.

This commit introduces a button within the Flickr component that dispatches an action to call the Flickr API and update the cached data.
A feedback message is displayed to inform the user about the ongoing process.

The global nav images are also updated.

* Adds Flickr latest photos component

Implements a new component to display the latest photos from Flickr, fetching data from Firebase.

Introduces actions to manage the latest photos data, including setting the photos array and the current photo index, and removes the reload button.
Now the app displays the latest photo from the flickr API.